Fix PC and MAR underflow, reformat code, ...
This commit fixes the incorrect increment/decrement of both the program
counter and the MAR when shell commands are run.

This commit also reformats the code in `src/main.c` to comply with new
style guidelines (i.e. C-style commenting, 4-space indents, 80
character widths, etc.).

This commit also fixes the build system in `Makefile` to properly clean
the `bin/` directory.